Fix giant bug: crash when don't config all modules
that's bacause you import all the modules
> from .middleware import github, keybase, medium
while each module need to get configurations from modules.json, e.g.
> config = Config.modules['modules']['github']
but can't get anything at all, so it will crash.

that's not the correct behaviour, so we just import the desired module
and then register it. The solution isn't very elegant.